rowexec: inline column backfiller implementation
This commit removes `chunkBackfiller` interface since it only has
a single implementation nowadays and inlines it into the column
backfiller. This allows us to simplify the code a bit.

This commit mostly a no-op change - the only material difference is the
removal of `b.chunks.CurrentBufferFill() > opportunisticCheckpointThreshold`
condition which was never true given the method always returned 0.

Release note: None
